Reorganize `polling` module
